mark-update fix

+++ public/js/delete-user.js
@@ -11,7 +11,6 @@ function confirmDelete() {
             }
         }).then(response => {
             if (response.ok) {
-                // Redirect to login page after successful deletion
                 window.location.href = '/login';
             } else {
                 alert('Error deleting account.');
@@ -26,22 +25,33 @@ document.addEventListener('DOMContentLoaded', () => {
     // Add event listener for delete buttons
     document.querySelectorAll('.danger.icon').forEach(button => {
         button.addEventListener('click', async (event) => {
-            const userId = event.target.dataset.id;
+            const userId = event.currentTarget.dataset.id; 
 
             try {
-                const response = await fetch(`/admin/delete-user/${userId}`, {
-                    method: 'DELETE'
-                });
+                if (confirm("Are you sure you want to delete your account? This action cannot be undone.")) {
+                    const response = await fetch(`/admin/delete-user/${userId}`, {
+                        method: 'DELETE'
+                    });
 
-                if (!response.ok) {
-                    throw new Error('Failed to delete user');
+                    if (!response.ok) {
+                        throw new Error('Failed to delete user');
+                    }
+                    
+                    // Refresh the page after successful deletion
+                    window.location.reload();
                 }
-
-                // Refresh the page after successful deletion
-                window.location.reload();
             } catch (error) {
                 console.error('Error deleting user:', error);
             }
+        });
+    });
+
+    document.querySelectorAll('.edit.icon').forEach(button => {
+        button.addEventListener('click', (event) => {
+            const userId = event.currentTarget.dataset.id;
+
+            // Redirect to the edit user page
+            window.location.href = `/admin/edit-user/${userId}`;
         });
     });
 });
